requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.4
django-admin-honeypot==1.1.0
django-allauth==0.42.0
django-progressive-web-app==0.1.1
gunicorn==20.0.4
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.4
python3-openid==3.2.0
pytz==2019.3
requests==2.24.0
requests-oauthlib==1.3.0
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.25.10
whitenoise==5.0.1
django-admin-csvexport==1.11